add class component
